Create minecraft.conf
1.12.2 Specific minecraft.conf

`block-fertilize` - Source needs to be `minecraft:dye` as bonemeal is a dye in 1.12.2

`falling-block-break` - Does not work, there is no source showing in a debug when a block falls on 1.12.2. - Removed

`fire-entity-damage` - Source `magma_block` is incorrect , is just `magma` on 1.12.2. Also was missing `hot_floor` as source

ravager-block-break doesn't exist in 1.12.2, - Removed

turtle-egg-hatch doesn't exist in 1.12.2, - Removed.